HBASE-19239 Fix findbugs and error-prone issues
Fixes for hbase-server

+++ hbase-server/src/test/java/org/apache/hadoop/hbase/http/HttpServerFunctionalTest.java
@@ -242,12 +242,15 @@ public class HttpServerFunctionalTest extends Assert {
    */
   protected static void deleteRecursively(File d) {
     if (d.isDirectory()) {
-      for (String name : d.list()) {
-        File child = new File(d, name);
-        if (child.isFile()) {
-          child.delete();
-        } else {
-          deleteRecursively(child);
+      String[] dirs = d.list();
+      if (dirs != null) {
+        for (String name: dirs) {
+          File child = new File(d, name);
+          if (child.isFile()) {
+            child.delete();
+          } else {
+            deleteRecursively(child);
+          }
         }
       }
     }
